Sport_notification/requirements.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1
djangorestframework==3.12.2
idna==2.10
parameterized==0.8.1
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
schedule==1.0.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
